requirements.txt
docker==5.0.0
six==1.16.0
pushbullet.py==0.12.0
